+++ crates/proto/build.rs
@@ -7,7 +7,7 @@ use miden_node_proto_build::{
     remote_prover_api_descriptor,
     rpc_api_descriptor,
     store_block_producer_api_descriptor,
-    store_ntx_builder_api_descriptor,
+    store_ntx_producer_api_descriptor,
     store_rpc_api_descriptor,
     validator_api_descriptor,
 };
@@ -41,7 +41,7 @@ fn main() -> miette::Result<()> {
 
     generate_bindings(rpc_api_descriptor(), &dst_dir)?;
     generate_bindings(store_rpc_api_descriptor(), &dst_dir)?;
-    generate_bindings(store_ntx_builder_api_descriptor(), &dst_dir)?;
+    generate_bindings(store_ntx_producer_api_descriptor(), &dst_dir)?;
     generate_bindings(store_block_producer_api_descriptor(), &dst_dir)?;
     generate_bindings(block_producer_api_descriptor(), &dst_dir)?;
     generate_bindings(remote_prover_api_descriptor(), &dst_dir)?;