+++ crates/proto/src/clients/mod.rs
@@ -6,19 +6,19 @@
 //! # Examples
 //!
 //! ```rust
-//! # use miden_node_proto::clients::{Builder, WantsTls, StoreNtxBuilderClient};
+//! # use miden_node_proto::clients::{Builder, WantsTls, StoreNtxProducerClient};
 //! # use url::Url;
 //!
 //! # async fn example() -> anyhow::Result<()> {
 //! // Create a store client with OTEL and TLS
 //! let url = Url::parse("https://example.com:8080")?;
-//! let client: StoreNtxBuilderClient = Builder::new(url)
+//! let client: StoreNtxProducerClient = Builder::new(url)
 //!     .with_tls()?                   // or `.without_tls()`
 //!     .without_timeout()             // or `.with_timeout(Duration::from_secs(10))`
 //!     .without_metadata_version()    // or `.with_metadata_version("1.0".into())`
 //!     .without_metadata_genesis()    // or `.with_metadata_genesis(genesis)`
 //!     .with_otel_context_injection() // or `.without_otel_context_injection()`
-//!     .connect::<StoreNtxBuilderClient>()
+//!     .connect::<StoreNtxProducerClient>()
 //!     .await?;
 //! # Ok(())
 //! # }
@@ -111,8 +111,8 @@ type InterceptedChannel = InterceptedService<Channel, Interceptor>;
 type GeneratedRpcClient = generated::rpc::api_client::ApiClient<InterceptedChannel>;
 type GeneratedBlockProducerClient =
     generated::block_producer::api_client::ApiClient<InterceptedChannel>;
-type GeneratedStoreClientForNtxBuilder =
-    generated::store::ntx_builder_client::NtxBuilderClient<InterceptedChannel>;
+type GeneratedStoreClientForNtxProducer =
+    generated::store::ntx_producer_client::NtxProducerClient<InterceptedChannel>;
 type GeneratedStoreClientForBlockProducer =
     generated::store::block_producer_client::BlockProducerClient<InterceptedChannel>;
 type GeneratedStoreClientForRpc = generated::store::rpc_client::RpcClient<InterceptedChannel>;
@@ -129,7 +129,7 @@ pub struct RpcClient(GeneratedRpcClient);
 #[derive(Debug, Clone)]
 pub struct BlockProducerClient(GeneratedBlockProducerClient);
 #[derive(Debug, Clone)]
-pub struct StoreNtxBuilderClient(GeneratedStoreClientForNtxBuilder);
+pub struct StoreNtxProducerClient(GeneratedStoreClientForNtxProducer);
 #[derive(Debug, Clone)]
 pub struct StoreBlockProducerClient(GeneratedStoreClientForBlockProducer);
 #[derive(Debug, Clone)]
@@ -169,14 +169,14 @@ impl Deref for BlockProducerClient {
     }
 }
 
-impl DerefMut for StoreNtxBuilderClient {
+impl DerefMut for StoreNtxProducerClient {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.0
     }
 }
 
-impl Deref for StoreNtxBuilderClient {
-    type Target = GeneratedStoreClientForNtxBuilder;
+impl Deref for StoreNtxProducerClient {
+    type Target = GeneratedStoreClientForNtxProducer;
 
     fn deref(&self) -> &Self::Target {
         &self.0
@@ -273,9 +273,9 @@ impl GrpcClient for BlockProducerClient {
     }
 }
 
-impl GrpcClient for StoreNtxBuilderClient {
+impl GrpcClient for StoreNtxProducerClient {
     fn with_interceptor(channel: Channel, interceptor: Interceptor) -> Self {
-        Self(GeneratedStoreClientForNtxBuilder::new(InterceptedService::new(
+        Self(GeneratedStoreClientForNtxProducer::new(InterceptedService::new(
             channel,
             interceptor,
         )))